refactor(monorepo): extract @chm/mcp-server + run package tests in CI (Phase 4c)
Extract @chm/mcp-server (lib/mcp + auth/api-key + auth/bearer-token +
data/mcp-tools-data) and move the dependency-free SQL validator into
@chm/sql-builder. Removes the apps/mcp-worker include-hack: the worker now
imports `@chm/mcp-server` instead of reaching into `../../web/...`.

- auth/data consumers use @chm/mcp-server/auth and /data subpaths so the MCP
  SDK never enters the Edge Runtime (middleware) bundle; the worker + MCP route
  use the full barrel/server export.
- fix tests left referencing moved files: integration-environment.test +
  api-key.test moved into their packages (relative paths resolve); fetchdata-
  hostid meta-test repointed to packages/clickhouse-client.
- CI: add a "Run workspace package tests" step (bun test packages) so the
  extracted package tests actually run (they're outside apps/web's suite).

Verified: type-check 0 errors, next build (standalone), wrangler dry-run
768 KiB, apps/web 353 tests pass, packages 289 tests pass.
