+++ apps/mcp-worker/package.json
@@ -9,6 +9,7 @@
     "typegen": "wrangler types"
   },
   "dependencies": {
+    "@chm/mcp-server": "workspace:*",
     "@modelcontextprotocol/sdk": "^1.29.0"
   },
   "devDependencies": {
