+++ apps/web/lib/ai/agent/__tests__/mcp-tool-adapter.test.ts
@@ -4,7 +4,7 @@
  * Imports from ../tools directly to avoid mock leaks from clickhouse-agent.test.ts
  * which mocks @/lib/ai/agent/mcp-tool-adapter globally.
  *
- * Does NOT mock @/lib/api/shared/validators/sql to avoid leaking into SQL validator tests.
+ * Does NOT mock @chm/sql-builder to avoid leaking into SQL validator tests.
  * Tool execute tests use readOnlyQuery (no validation) so the real validator isn't invoked.
  */
