+++ apps/web/lib/api/shared/index.ts
@@ -34,5 +34,4 @@ export {
   validateHostIdWithError,
   validateRequiredString,
   validateSearchParams,
-  validateSqlQuery,
 } from './validators'
