+++ apps/web/package.json
@@ -63,6 +63,7 @@
     "@aws-sdk/xml-builder": "^3.972.26",
     "@chm/clickhouse-client": "workspace:*",
     "@chm/logger": "workspace:*",
+    "@chm/mcp-server": "workspace:*",
     "@chm/platform": "workspace:*",
     "@chm/sql-builder": "workspace:*",
     "@chm/types": "workspace:*",
