+++ packages/mcp-server/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "@chm/mcp-server",
+  "version": "0.1.0",
+  "private": false,
+  "type": "module",
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts",
+    "./server": "./src/server.ts",
+    "./auth": "./src/auth/index.ts",
+    "./data": "./src/data/mcp-tools-data.ts"
+  },
+  "dependencies": {
+    "@chm/clickhouse-client": "workspace:*",
+    "@chm/sql-builder": "workspace:*",
+    "@clickhouse/client": "^1.18.5",
+    "@modelcontextprotocol/sdk": "^1.29.0",
+    "zod": "^4.4.3"
+  },
+  "devDependencies": {
+    "typescript": "^6.0.0"
+  }
+}
