+++ packages/mcp-server/src/__tests__/performance.test.ts
