+++ packages/mcp-server/src/__tests__/prompts.test.ts
