+++ packages/mcp-server/src/__tests__/server.test.ts
