+++ packages/mcp-server/src/__tests__/tools.test.ts
