+++ packages/mcp-server/src/auth/api-key.ts
