+++ packages/mcp-server/src/auth/bearer-token.ts
