+++ packages/mcp-server/src/prompts/index.ts
