+++ packages/mcp-server/src/resources/index.ts
