+++ packages/mcp-server/src/server.ts
