+++ packages/mcp-server/src/tools/databases.ts
