+++ packages/mcp-server/src/tools/explore-table-schema.ts
