+++ packages/mcp-server/src/tools/index.ts
