+++ packages/mcp-server/src/tools/performance.ts
