+++ packages/mcp-server/src/tools/query.ts
@@ -3,8 +3,8 @@ import type { DataFormat } from '@clickhouse/client'
 import type { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js'
 
 import { fetchData } from '@chm/clickhouse-client'
+import { validateSqlQuery } from '@chm/sql-builder'
 import { z } from 'zod/v3'
-import { validateSqlQuery } from '@/lib/api/shared/validators/sql'
 
 export function registerQueryTool(server: McpServer) {
   server.tool(
