+++ packages/mcp-server/src/tools/tables.ts
