+++ packages/sql-builder/src/index.ts
@@ -44,5 +44,6 @@ export { ExtendedBuilder } from './extension'
 export { fn } from './functions'
 export { getAllSqlStrings } from './query-config-types'
 export { param, RawSql, raw } from './raw'
+export { SQL_PATTERNS, validateSqlQuery } from './sql-validator'
 export * from './types'
 export { SqlBuilderError, validateBuilderState } from './validator'
